Rename viml, lua, vim_compatible -> lush_to_viml, lush_to_lua, vim_compatible_viml
