+++ spec/transform/lua_spec.moon
@@ -2,6 +2,7 @@ describe "transforms run.lua", ->
   run = require("lush.builder").run
 
   setup ->
+    lua = require("lush.transform.lush_to_lua")
     parse = require('lush.parser')
     ast = parse -> {
       A { gui: "italic", blend: 40 },
@@ -19,6 +20,5 @@ describe "transforms run.lua", ->
     package.loaded["theme"] = nil
 
   it "returns lua code", ->
-    lua = require("lush.transform.lua")
     value = run(require("theme"), lua)
     assert.is.table(value)
